se cambio precio de producto a double para permitir registrar recetas con fraccioens

+++ src/main/java/lacosmetics/planta/lacmanufacture/model/producto/Producto.java
@@ -43,7 +43,7 @@ public abstract class Producto {
     private String observaciones;
 
     @Min(value=0, message = "El costo no puede ser negativo") // validacion en db engine
-    private int costo;
+    private double costo;
 
     /**
      * valores vigentes: 0%, 5% y 19%
